refactor WarcWriterThread and WarcPerUrlWriterThread into seperate warcwriters module. Add --warc-per-url option to use the WarcPerUrlWriterThread
